fix: properly sync all unit types on live map
