+++ apps/client/src/components/dispatch/map/units/active-map-units.tsx
@@ -1,12 +1,5 @@
 import * as React from "react";
-import {
-  CombinedEmsFdUnit,
-  CombinedLeoUnit,
-  EmsFdDeputy,
-  Officer,
-  ShouldDoType,
-} from "@snailycad/types";
-import { isUnitCombined, isUnitCombinedEmsFd } from "@snailycad/utils";
+import { CombinedEmsFdUnit, CombinedLeoUnit, EmsFdDeputy, Officer } from "@snailycad/types";
 import { useActiveDeputies } from "hooks/realtime/useActiveDeputies";
 import { useActiveOfficers } from "hooks/realtime/useActiveOfficers";
 import type { MapPlayer, PlayerDataEventPayload } from "types/map";
@@ -17,6 +10,7 @@ import { useTranslations } from "next-intl";
 import { UnitItem } from "./unit-item";
 import { ManageUnitModal } from "components/dispatch/modals/manage-unit-modal";
 import { useMapPlayersStore } from "hooks/realtime/use-map-players";
+import { createMapUnitsFromActiveUnits } from "lib/map/create-map-units-from-active-units.ts";
 
 interface Props {
   players: (MapPlayer | PlayerDataEventPayload)[];
@@ -35,7 +29,7 @@ export function ActiveMapUnits({ openItems, setOpenItems }: Props) {
 
   const { activeOfficers } = useActiveOfficers();
   const { activeDeputies } = useActiveDeputies();
-  const units = makeActiveUnits({
+  const units = createMapUnitsFromActiveUnits({
     players: Array.from(players.values()),
     activeOfficers,
     activeDeputies,
@@ -69,77 +63,5 @@ export function ActiveMapUnits({ openItems, setOpenItems }: Props) {
       </div>,
       portalRef,
     )
-  );
-}
-
-interface ActiveUnitsOptions {
-  players: (MapPlayer | PlayerDataEventPayload)[];
-  activeOfficers: (Officer | CombinedLeoUnit)[];
-  activeDeputies: (EmsFdDeputy | CombinedEmsFdUnit)[];
-}
-
-function makeActiveUnits({ players, activeOfficers, activeDeputies }: ActiveUnitsOptions) {
-  const activeUnits: MapPlayer[] = [];
-  const _activeOfficers = activeOfficers;
-  const _activeDeputies = activeDeputies;
-
-  for (const activeUnit of [..._activeOfficers, ..._activeDeputies]) {
-    if (!activeUnit.status || activeUnit.status.shouldDo === ShouldDoType.SET_OFF_DUTY) continue;
-
-    if (isUnitCombined(activeUnit) || isUnitCombinedEmsFd(activeUnit)) {
-      const player = players.find((player) => findPlayer(player, activeUnit));
-
-      if (!player || !("steamId" in player) || !("discordId" in player)) continue;
-
-      const existing = activeUnits.some((player) => findPlayer(player, activeUnit));
-
-      if (player && !existing) {
-        activeUnits.push({ ...player, unit: activeUnit });
-      }
-
-      continue;
-    }
-
-    if (!activeUnit.user) continue;
-
-    const steamId = activeUnit.user.steamId;
-    const discordId = activeUnit.user.discordId;
-
-    const player = players.find((player) => {
-      return player.discordId === discordId || player.convertedSteamId === steamId;
-    });
-
-    if (!player || !("steamId" in player) || !("discordId" in player)) continue;
-
-    const existing = activeUnits.some((unit) => {
-      return unit.discordId === discordId || unit.convertedSteamId === steamId;
-    });
-
-    if (player && !existing) {
-      activeUnits.push({ ...player, unit: activeUnit });
-    }
-  }
-
-  return activeUnits;
-}
-
-function findPlayer(
-  player: MapPlayer | PlayerDataEventPayload,
-  activeUnit: CombinedLeoUnit | CombinedEmsFdUnit,
-) {
-  const steamIds = isUnitCombinedEmsFd(activeUnit)
-    ? activeUnit.deputies.map((deputy) => deputy.user?.steamId)
-    : activeUnit.officers.map((officer) => officer.user?.steamId);
-
-  const discordIds = isUnitCombinedEmsFd(activeUnit)
-    ? activeUnit.deputies.map((deputy) => deputy.user?.discordId)
-    : activeUnit.officers.map((officer) => officer.user?.discordId);
-
-  const filteredSteamIds = steamIds.filter(Boolean) as string[];
-  const filteredDiscordIds = discordIds.filter(Boolean) as string[];
-
-  return (
-    (player.convertedSteamId && filteredSteamIds.includes(player.convertedSteamId)) ||
-    (player.discordId && filteredDiscordIds.includes(player.discordId))
   );
 }
